Clean up some test code...
* Some builtin function tests moved to their more specific location
* isort imports
* skipped tests are marked skipped rather than expected fail (becuase we
  haven't figured out how to do something, rather than there being an
  inherent problem
