+++ test/builtin/numbers/test_linalg.py
@@ -1,8 +1,9 @@
 # -*- coding: utf-8 -*-
 """
-Unit tests from mathics.builtins.numbers.linalg
+Unit tests for mathics.builtins.numbers.linalg
 """
 from test.helper import check_evaluation
+
 import pytest
