+++ test/builtin/numbers/test_nintegrate.py
@@ -1,13 +1,15 @@
 # -*- coding: utf-8 -*-
 """
+Unit tests for mathics.buitin.numbers.nintegrate
+
 NIntegrate[] tests
 
 """
-import importlib
-import pytest
 from test.helper import evaluate
-from mathics.builtin.base import check_requires_list
 
+import pytest
+
+from mathics.builtin.base import check_requires_list
 
 if check_requires_list(["scipy", "scipy.integrate"]):
     methods = ["Automatic", "Romberg", "Internal", "NQuadrature"]
